file i/o

+++ 0x15-file_io/0-read_textfile.c
@@ -22,7 +22,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	return (0);
 
 	ch_read = read(read_textfile, buffer, letters);
-	if (ch_read == -1)
+	if (ch_read == (size_t)-1)
 	return (0);
 
 	ch_write = write(STDOUT_FILENO, buffer, ch_read);
